added conversion to SEM

+++ src/path_utils.py
@@ -2,7 +2,10 @@ import os
 import json
 import shutil
 
-def get_relevant_info(repo_dir, project, verbose=0):
+def get_relevant_info(repo_dir,
+                      project,
+                      start_from_scratch=False,
+                      verbose=0):
     """
 
     :param str repo_dir: the absolute path of the DFNDataReleases repository,
@@ -21,6 +24,13 @@ def get_relevant_info(repo_dir, project, verbose=0):
     for name, path in paths.items():
         paths[name] = json.load(open(path))
 
+    inc2type = {}
+    for type_, incs in paths['type2inc'].items():
+        for inc in incs:
+            assert inc not in inc2type, f'{inc} has been assigned to 2> event types. Please inspect.'
+            inc2type[inc] = type_
+    paths['inc2type'] = inc2type
+
     paths['inc_coll_obj'] = os.path.join(repo_dir, 'structured', 'data_release_inc_coll_obj.p')
     paths['unstructured'] = os.path.join(repo_dir, 'unstructured')
     paths['main_statistics_folder'] = os.path.join(repo_dir, 'statistics')
@@ -33,12 +43,15 @@ def get_relevant_info(repo_dir, project, verbose=0):
 
     project_dir = paths['project_statistics']
     if os.path.exists(project_dir):
-        shutil.rmtree(project_dir)
+        if start_from_scratch:
+            shutil.rmtree(project_dir)
+            if verbose >= 1:
+                print(f'removed existing folder {project_dir}')
+
+    if not os.path.exists(project_dir):
+        os.mkdir(project_dir)
         if verbose >= 1:
-            print(f'removed existing folder {project_dir}')
-    os.mkdir(project_dir)
-    if verbose >= 1:
-        print(f'created folder at {project_dir}')
+            print(f'created folder at {project_dir}')
 
     for name, path in paths.items():
         assert os.path.exists, f'{path} does not exist'
